\name{temp}
\alias{plot.umlarrow}
\alias{umlarrow}
\alias{plot.umlbox}
\alias{setbox}
\alias{umlbounds.umlbox}
\alias{umlbox}
\alias{umlclass}
\alias{umlgui}
\alias{plot.umlgraph}
\alias{plot.umlnode}
\alias{umlbounds}
\alias{umlbounds.umlnode}
\alias{umlcanvas}
\alias{umlgraph}
\alias{umlgraph.pdf}
\alias{umlgraph.preview}
\alias{umlnode}
\title{Temporary Documentation File}
\description{Documenting pending, next revision...}
\usage{
umlarrow (v1, v2)
setbox (k, x, y)
umlbox (str=list(""), col=NA, border=black, x=0, y=0, wmin=5, vincr=0.25)
umlclass (name, attributes=NULL, methods=NULL, \dots)
umlgui (g)
umlbounds (v, \dots)
umlcanvas (xlim=c(-10, 10), ylim=c(10, -10), gridlines=TRUE)
umlgraph (vs, cs)
umlgraph.pdf (g)
umlgraph.preview (g)
umlnode (x, y)
\method{plot}{umlarrow}(con, \dots)
\method{plot}{umlbox}(k, \dots)
\method{umlbounds}{umlbox}(v, \dots)
\method{plot}{umlgraph}(g, \dots)
\method{plot}{umlnode}(v, \dots)
\method{umlbounds}{umlnode}(v, \dots)
}
\arguments{
\item{x}{.}
\item{y}{.}
\item{v}{.}
\item{g}{.}
\item{k}{.}
\item{v1}{.}
\item{v2}{.}
\item{vs}{.}
\item{cs}{.}
\item{con}{.}
\item{str}{.}
\item{col}{.}
\item{border}{.}
\item{xlim}{.}
\item{ylim}{.}
\item{wmin}{.}
\item{vincr}{.}
\item{name}{.}
\item{attributes}{.}
\item{methods}{.}
\item{gridlines}{.}
\item{\dots}{.}
}

