\name{umlarrow}
\alias{umlarrow}
\alias{umlextends}
\alias{plot.umlarrow}
\title{uml arrows}
\description{todo}
\usage{
umlarrow (v1, v2, solid=FALSE, col="black", fill=getOption("umlr.fill") )
umlextends (v1, v2, \dots)
\method{plot}{umlarrow}(con, \dots)
}
\arguments{
\item{v1}{.}
\item{v2}{.}
\item{con}{.}
\item{solid}{.}
\item{col}{.}
\item{fill}{.}
\item{\dots}{.}
}



