\name{umlbox}
\alias{umlbox}
\alias{umldims.umlbox}
\alias{plot.umlbox}
\title{uml boxes}
\description{todo}
\usage{
umlbox (x=0, y=0, w=getOption ("umlr.wmin"), h=1.2, text="",
	col="black", fill=getOption("umlr.fill") )
\method{umldims}{umlbox}(v, \dots)
\method{plot}{umlbox}(v, \dots)
}
\arguments{
\item{x}{.}
\item{y}{.}
\item{w}{.}
\item{h}{.}
\item{text}{.}
\item{col}{.}
\item{fill}{.}
\item{v}{.}
\item{\dots}{.}
}


