\name{umlgraph}
\alias{umlgraph}
\alias{umlnode}
\alias{umlconnection}
\alias{plot.umlgraph}
\alias{plot.umlnode}
\alias{plot.umlconnection}
\alias{umldims.umlgraph}
\alias{umldims.umlnode}
\alias{recenter.umlnode}
\title{uml graphs}
\description{todo}
\usage{
umlgraph (vs=list (), cs=list () )
umlnode (x=0, y=0)
umlconnection (v1, v2)
\method{plot}{umlgraph}(g, \dots)
\method{plot}{umlnode}(v, \dots)
\method{plot}{umlconnection}(con, \dots)
\method{umldims}{umlgraph}(g, \dots)
\method{umldims}{umlnode}(v, \dots)
\method{recenter}{umlnode}(v, x, y, \dots)
}
\arguments{
\item{vs}{.}
\item{cs}{.}
\item{x}{.}
\item{y}{.}
\item{v1}{.}
\item{v2}{.}
\item{g}{.}
\item{v}{.}
\item{con}{.}
\item{\dots}{.}
}


