\name{umlplot}
\alias{plot.uml}
\alias{umlpdf}
\alias{umlpng}
\alias{plot.umlcanvas}
\title{uml plots}
\description{todo}
\usage{
\method{plot}{uml}(m, fresh=TRUE, crop=FALSE, grid=FALSE, \dots)
umlpdf (m, image.path=paste (m$image.path, "pdf", sep=".") )
umlpng (m, image.path=paste (m$image.path, "png", sep=".") )
\method{plot}{umlcanvas}(k, grid=FALSE, \dots)
}
\arguments{
\item{m}{.}
\item{fresh}{.}
\item{crop}{.}
\item{grid}{.}
\item{image.path}{.}
\item{k}{.}
\item{\dots}{.}
}


