% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umlr.R
\docType{package}
\name{umlr-package}
\alias{umlr}
\alias{umlr-package}
\title{Helper functions to create UML diagrams from R code}
\description{
Helper functions to create UML diagrams from R code.
}
\details{
Details
}
\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[umlr]{render}}: }{

     Renders UML diagrams
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state.}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/umlr}
}
\seealso{
See \code{vignette("introduction", package = "umlr")}.
}

