\name{uml}
\alias{uml}
\alias{umlcanvas}
\alias{print.uml}
\alias{umlsave}
\alias{umlload}
\alias{umldims}
\alias{umldims.uml}
\alias{umlro.fill}
\alias{recenter}
\title{uml models}
\description{todo}
\usage{
uml (\dots, g=umlgraph (...), k=umlcanvas (), model.path="model.umlr", image.path="model")
umlcanvas (model.size=c (22, 15), margin.size=0.2, xlim, ylim)
\method{print}{uml}(m, \dots)
umlsave (m, path=m$model.path)
umlload (path="model.umlr")
umldims (\dots)
\method{umldims}{uml}(m, \dots)
umlro.fill (fill="white")
recenter (\dots)
}
\arguments{
\item{g}{.}
\item{k}{.}
\item{model.path}{.}
\item{image.path}{.}
\item{model.size}{.}
\item{margin.size}{.}
\item{xlim}{.}
\item{ylim}{.}
\item{m}{.}
\item{path}{.}
\item{fill}{.}
\item{\dots}{.}
}



