\name{umlcomplex}
\alias{umlcomplex}
\alias{print.umlcomplex}
\alias{plot.umlcomplex}
\alias{umlclass}
\title{uml boxes with partitions}
\description{todo}
\usage{
umlcomplex (x=0, y=0, text=list (""), markup=FALSE, italics=FALSE,
	expand=FALSE, roundness=0.18, \dots)
\method{print}{umlcomplex}(v, \dots)
\method{plot}{umlcomplex}(v, \dots)
umlclass (name, attributes=NULL, methods=NULL, abstract=FALSE, \dots)
}
\arguments{
\item{name}{.}
\item{attributes}{.}
\item{methods}{.}
\item{abstract}{.}
\item{x}{.}
\item{y}{.}
\item{text}{.}
\item{markup}{.}
\item{italics}{.}
\item{expand}{.}
\item{roundness}{.}
\item{v}{.}
\item{\dots}{.}
}




