\name{umlgraph}
\alias{umlgraph}
\alias{umlnode}
\alias{umlconnection}
\alias{is.umlnode}
\alias{is.umlconnection}
\alias{print.umlgraph}
\alias{print.umlnode}
\alias{print.umlconnection}
\alias{plot.umlgraph}
\alias{plot.umlnode}
\alias{plot.umlconnection}
\alias{umldims.umlgraph}
\alias{umldims.umlnode}
\alias{recenter.umlnode}
\alias{northc}
\title{uml graphs}
\description{todo}
\usage{
umlgraph (\dots, vs=list (), cs=list () )
umlnode (x=0, y=0, snapf=northc)
umlconnection (v1, v2)
\method{print}{umlgraph}(g, \dots)
\method{print}{umlnode}(v, \dots)
\method{print}{umlconnection}(con, \dots)
\method{plot}{umlgraph}(g, \dots)
\method{plot}{umlnode}(v, \dots)
\method{plot}{umlconnection}(con, \dots)
is.umlnode (obj)
is.umlconnection (obj)
\method{umldims}{umlgraph}(g, \dots)
\method{umldims}{umlnode}(v, \dots)
\method{recenter}{umlnode}(v, x, y, \dots)
northc (v, x, y)
}
\arguments{
\item{obj}{.}
\item{vs}{.}
\item{cs}{.}
\item{x}{.}
\item{y}{.}
\item{snapf}{.}
\item{v1}{.}
\item{v2}{.}
\item{g}{.}
\item{v}{.}
\item{con}{.}
\item{\dots}{.}
}




